# geoFields/src/main/resources/static/OpenAPIspec.yaml
openapi: 3.0.3
info:
  title: GeoFields HTTP API
  description: |
    JSON-эндпоинты бэкенда GeoFields (поля, NDVI, менеджер организации).

    **Авторизация:** сессия Spring Security (форма логина). Запросы с браузера отправляйте с `credentials: include` / same-origin cookie.

    **CSRF:** для `POST` используется cookie `XSRF-TOKEN` и заголовок `X-XSRF-TOKEN` (см. Spring `CookieCsrfTokenRepository`).

    Главная `/` отдаёт HTML без данных пользователя; контекст сессии — `GET /api/session/context`.

    HTML-страницы (`/`, `/login`, `/register`, `/org/manager`, `/org/admin`, `/org/agronomist`) как разметка в спецификации не описаны.
  version: 1.4.3
servers:
  - url: http://localhost:8080
    description: Локальный сервер (см. `server.port` в application.properties)
...
paths:
  /python/get_ndvi_tiles_for_field:
    get:
      tags: [PythonNDVI]
      summary: (Frontend → Python) URL тайлов NDVI для поля
      description: |
        Этот запрос выполняется **из браузера** (см. `static/js/app.js`) напрямую в Python API:
        `GET http://localhost:8001/get_ndvi_tiles_for_field?field_id=...&date=YYYY-MM-DD`.
      servers:
        - url: http://localhost:8001
          description: Python NDVI API
      parameters:
        - name: field_id
          in: query
          required: true
          schema: { type: string }
          example: "12"
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
          example: "2024-06-15"
      responses:
        "200":
          description: URL тайлов и фактическая дата снимка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PythonNdviTilesForFieldResponse"
        "404":
          description: Нет доступных спутниковых данных
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PythonErrorResponse"
        "500":
          description: Ошибка Python-сервиса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PythonErrorResponse"
  /python/get_ndvi_trend:
    get:
      tags: [PythonNDVI]
      summary: (Frontend → Python) Тренд NDVI по группе полей
      description: |
        Этот запрос выполняется **из браузера** напрямую в Python API:
        `GET http://localhost:8001/get_ndvi_trend?field_id=1&field_id=2&start_date=...&end_date=...`.
      servers:
        - url: http://localhost:8001
          description: Python NDVI API
      parameters:
        - name: field_id
          in: query
          required: true
          description: Идентификаторы полей (повторяющийся query-параметр)
          schema:
            type: array
            items: { type: integer }
          style: form
          explode: true
          example: [12, 15]
        - name: start_date
          in: query
          required: true
          schema: { type: string, format: date }
          example: "2024-04-01"
        - name: end_date
          in: query
          required: true
          schema: { type: string, format: date }
          example: "2024-09-01"
      responses:
        "200":
          description: Данные для построения графика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PythonNdviTrendResponse"
        "500":
          description: Ошибка Python-сервиса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PythonErrorResponse"
  /python/tiles/{field_id}/{date}/{scene_id}/{z}/{x}/{y}.png:
    get:
      tags: [PythonNDVI]
      summary: (MapLibre → Python) PNG-тайл NDVI
      description: |
        Тайловый URL используется картой как raster tiles (запросы формирует MapLibre GL).
        Реальный эндпоинт: `GET http://localhost:8001/tiles/{field_id}/{date}/{scene_id}/{z}/{x}/{y}.png`.
      servers:
        - url: http://localhost:8001
          description: Python NDVI API
      parameters:
        - name: field_id
          in: path
          required: true
          schema: { type: string }
          example: "12"
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
          example: "2024-06-15"
        - name: scene_id
          in: path
          required: true
          schema: { type: string }
          example: "S2A_20240615"
        - name: z
          in: path
          required: true
          schema: { type: integer }
          example: 10
        - name: x
          in: path
          required: true
          schema: { type: integer }
          example: 615
        - name: y
          in: path
          required: true
          schema: { type: integer }
          example: 342
      responses:
        "200":
          description: PNG изображение тайла
          content:
            image/png:
              schema:
                type: string
                format: binary
...
  /get_ndvi_tiles_for_field:
    get:
      tags: [NDVI]
      summary: URL тайлов NDVI для одного поля (Java proxy → Python)
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: query
          required: true
          schema:
            type: string
          example: "1"
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-06-15"
      responses:
        "200":
          description: URL тайлов и фактическая дата снимка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdviTilesForFieldResponse"
        "400":
          description: Некорректный field_id или дата
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
        "502":
          description: Python slope service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /get_ndvi_trend:
    get:
      tags: [NDVI]
      summary: Тренд NDVI по группе полей (Java proxy → Python)
      description: |
        Перед получением тренда Java может выполнить прогрев расчётов помесячно (`warmup=true`),
        чтобы уменьшить вероятность пустого графика при первом запросе.
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: query
          required: true
          description: Идентификаторы полей (повторяющийся query-параметр)
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
          example: [1, 2]
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2023-01-01"
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: warmup
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Включить помесячный прогрев NDVI перед запросом тренда
      responses:
        "200":
          description: Данные для графика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdviTrendResponse"
        "400":
          description: Некорректные параметры
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
        "502":
          description: Python slope service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /tiles/{field_id}/{date}/{scene_id}/{z}/{x}/{y}.png:
    get:
      tags: [NDVI]
      summary: PNG-тайл NDVI (Java proxy + cache)
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: path
          required: true
          schema: { type: string }
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
        - name: scene_id
          in: path
          required: true
          schema: { type: string }
        - name: z
          in: path
          required: true
          schema: { type: integer }
        - name: x
          in: path
          required: true
          schema: { type: integer }
        - name: y
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PNG тайл
          headers:
            X-Ndvi-Cache:
              schema:
                type: string
              description: Источник кэша (`MISS`, `HIT-MEMORY`, `HIT-S3`)
          content:
            image/png:
              schema:
                type: string
                format: binary
        "204":
          description: Пустой тайл (нет данных)
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
        "502":
          description: Python slope service unavailable
...
  /prepare_elevation/{field_id}:
    get:
      tags: [Elevation]
      summary: Подготовить рельеф и уклон (SRTM) для поля
      description: |
        Прокси к Python `/prepare_elevation/{field_id}`.
        Возвращает URL шаблоны тайлов рельефа и уклона относительно Java-сервера.
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: path
          required: true
          schema: { type: string }
          example: "1"
      responses:
        "200":
          description: Данные готовы к отображению на карте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepareElevationResponse"
        "400":
          description: Некорректный field_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
        "502":
          description: Python elevation service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /tiles/elevation/{field_id}/{z}/{x}/{y}.png:
    get:
      tags: [Elevation]
      summary: PNG-тайл рельефа (SRTM)
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: path
          required: true
          schema: { type: string }
        - name: z
          in: path
          required: true
          schema: { type: integer }
        - name: x
          in: path
          required: true
          schema: { type: integer }
        - name: y
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PNG тайл
          content:
            image/png:
              schema:
                type: string
                format: binary
        "204":
          description: Пустой тайл (нет данных)
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
        "502":
          description: Python elevation service unavailable
  /tiles/slope/{field_id}/{z}/{x}/{y}.png:
    get:
      tags: [Elevation]
      summary: PNG-тайл уклона по SRTM (не путать с /slope/tiles/… по дате снимка)
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: path
          required: true
          schema: { type: string }
        - name: z
          in: path
          required: true
          schema: { type: integer }
        - name: x
          in: path
          required: true
          schema: { type: integer }
        - name: y
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PNG тайл
          content:
            image/png:
              schema:
                type: string
                format: binary
        "204":
          description: Пустой тайл (нет данных)
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
        "502":
          description: Python elevation service unavailable
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Сессия после входа через POST /login
  schemas:
    PythonNdviTilesForFieldResponse:
      type: object
      properties:
        url:
          type: string
          description: URL шаблон тайлов `{z}/{x}/{y}`
          example: "http://localhost:8001/tiles/12/2024-06-15/S2A_20240615/{z}/{x}/{y}.png"
        actual_date:
          type: string
          format: date
          description: Фактическая дата снимка (может отличаться от запрошенной)
          example: "2024-06-15"
    PythonNdviTrendResponse:
      type: object
      properties:
        labels:
          type: array
          items: { type: string }
          example: ["Apr", "May", "Jun"]
        data:
          type: array
          items: { type: number, format: double }
          example: [0.41, 0.55, 0.62]
    NdviTilesForFieldResponse:
      type: object
      properties:
        url:
          type: string
          description: URL шаблон тайлов `{z}/{x}/{y}` через Java proxy (`/tiles/...`)
          example: "/tiles/12/2024-06-15/S2A_20240615/{z}/{x}/{y}.png"
        actual_date:
          type: string
          format: date
          description: Фактическая дата снимка (может отличаться от запрошенной)
          example: "2024-06-15"
    NdviTrendResponse:
      type: object
      properties:
        labels:
          type: array
          items: { type: string }
          example: ["Apr", "May", "Jun"]
        data:
          type: array
          items: { type: number, format: double }
          example: [0.41, 0.55, 0.62]
...
    PrepareElevationResponse:
      type: object
      required: [status, tile_url]
      properties:
        status:
          type: string
          example: ready
        tile_url:
          type: string
          description: URL шаблон тайлов рельефа `{z}/{x}/{y}`
          example: "/tiles/elevation/1/{z}/{x}/{y}.png"
        slope_tile_url:
          type: string
          description: URL шаблон тайлов уклона по SRTM
          example: "/tiles/slope/1/{z}/{x}/{y}.png"
        elevation_min:
          type: number
          format: double
          example: 112.5
        elevation_max:
          type: number
          format: double
          example: 186.3
    PythonErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Нет доступных спутниковых данных"
...
    SimpleError:
      type: object
      properties:
        error:
          type: string
          example: "Некорректный field_id"
...